requirements.txt
# M1：锁定版本以降低供应链漂移；升级时请重跑测试并对照安全公告。
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
SQLAlchemy==2.0.36
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.6
pydantic-settings==2.7.0
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1
